pyproject.toml
[project]
name = "agentic-day1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "amazon-textract-caller>=0.2.4",
    "amazon-textract-textractor>=1.9.2",
    "beautifulsoup4>=4.14.3",
    "boto3>=1.42.94",
    "botocore[crt]>=1.42.94",
    "colorama>=0.4.6",
    "deepagents>=0.5.3",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.1",
    "langchain-google-genai>=4.2.2",
    "langchain-openai>=1.2.0",
    "langchain-postgres>=0.0.17",
    "langchain-text-splitters>=1.1.2",
    "langgraph>=1.1.9",
    "lxml>=6.1.0",
    "matplotlib>=3.10.8",
    "openai>=2.32.0",
    "pdf2image>=1.17.0",
    "pdfplumber>=0.11.9",
    "pgvector>=0.3.6",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "pypdf>=6.10.2",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-magic>=0.4.27",
    "rank-bm25>=0.2.2",
    "reportlab>=4.4.10",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.4.1",
    "tavily-python>=0.7.23",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "flake8>=7.3.0",
    "mypy>=1.20.2",
    "pre-commit>=4.6.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.11",
]
